backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
django-cors-headers==4.5.0
djangorestframework==3.15.2
google-auth==2.27.0
google-auth-oauthlib==1.2.0
idna==3.10
mysqlclient==2.2.5
psycopg2==2.9.10
python-decouple==3.8
python-dotenv==1.0.1
python-dateutil==2.9.0.post0
requests==2.32.3
pytz==2025.2
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
